Auth

+++ TechnicalRadiation/Controllers/AuthorController.cs
@@ -1,4 +1,5 @@
 using Microsoft.AspNetCore.Mvc;
+using System.Net.Http;
 using Microsoft.Extensions.Configuration;
 using TechnicalRadiation.Models.InputModels;
 using TechnicalRadiation.Services;
@@ -22,7 +23,7 @@ namespace TechnicalRadiation.Controllers
             Config = config;
         }
 
-         // GET api/authors
+        // GET api/authors
         [Route("")]
         [HttpGet]
         public IActionResult GetAllAuthors([FromQuery] int pageSize, [FromQuery] int pageNumber)
@@ -46,11 +47,15 @@ namespace TechnicalRadiation.Controllers
             return Ok(_authorService.GetNewsByAuthorId(id));
         }
 
-         // Post api/authors
+        // Post api/authors
         [Route("")]
         [HttpPost]
-        public IActionResult CreateNewAuthor([FromBody] AuthorInputModel newAuthor, [FromHeader]string xApiKey)
+        public IActionResult CreateNewAuthor([FromBody] AuthorInputModel newAuthor, [FromHeader]string Authorization)
         {
+            if (Authentication.Authenticate(Authorization) == false)
+            {
+                return Unauthorized();
+            }
             if (!ModelState.IsValid)
             {
                 return BadRequest("Model is not properly formatted");
@@ -64,8 +69,12 @@ namespace TechnicalRadiation.Controllers
         // PUT api/authors/1
         [Route("{id:int}")]
         [HttpPut]
-        public IActionResult UpdateAuthorById( int id, [FromBody] AuthorInputModel author, [FromHeader]string xApiKey)
+        public IActionResult UpdateAuthorById(int id, [FromBody] AuthorInputModel author, [FromHeader]string Authorization)
         {
+            if (Authentication.Authenticate(Authorization) == false)
+            {
+                return Unauthorized();
+            }
             if (!ModelState.IsValid)
             {
                 return BadRequest("Model is not properly formatted");
@@ -77,17 +86,25 @@ namespace TechnicalRadiation.Controllers
         // DELETE api/authors/1
         [Route("{id:int}")]
         [HttpDelete]
-        public ActionResult DeleteAuthorById(int id, [FromHeader]string xApiKey)
+        public ActionResult DeleteAuthorById(int id, [FromHeader]string Authorization)
         {
+            if (Authentication.Authenticate(Authorization) == false)
+            {
+                return Unauthorized();
+            }
             _authorService.DeleteAuthorById(id);
             return NoContent();
         }
 
-         // Post api/authors/1/newsItems/1
+        // Post api/authors/1/newsItems/1
         [Route("{authorId:int}/newsItems/{newsItemId:int}")]
         [HttpPost]
-        public IActionResult LinkAuthorToNewsItem(int authorId, int newsItemId)
+        public IActionResult LinkAuthorToNewsItem(int authorId, int newsItemId, [FromHeader]string Authorization)
         {
+            if (Authentication.Authenticate(Authorization) == false)
+            {
+                return Unauthorized();
+            }
             _authorService.LinkAuthorToNewsItem(authorId, newsItemId);
             return Ok();
         }
